refactor: abstract code and inject deps
